altera o cadastro de transacoes para diferenciar parcela de trasacano unica

+++ src/main/java/br/com/mywallet/app/domain/model/Dashboard/GastosPorCategoriaDTO.java
@@ -1,6 +1,8 @@
 package br.com.mywallet.app.domain.model.Dashboard;
 
+import java.math.BigDecimal;
+
 public record GastosPorCategoriaDTO(
         String categoria,
-        Double valorTotal
+        BigDecimal valorTotal
 ) {}
